services/ingestor/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
